requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--no-binary psycopg2

arrow==0.13.0
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.2.1
attrs==18.2.0
babel==2.6.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==7.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
cryptography==2.5
daemonize==2.5.0
database-sanitizer==0.4.0
defusedxml==0.5.0
delorean==1.0.0
django-admin-json-editor==0.1.5
django-allauth==0.38.0
django-anymail==5.0
django-appconf==1.0.2
django-cors-headers==2.4.0
django-enumfields==1.0.0
django-environ==0.4.5
django-filter==2.1.0
django-grappelli==2.12.2
django-guardian==1.4.9
django-helusers==0.4.2
django-hstore==1.4.2
django-image-cropping==1.2.0
django-jinja==2.4.1
django-js-asset==1.1.0
django-modeltranslation==0.13b1
django-mptt==0.9.1
django-multi-email-field==0.5.1
django-munigeo==0.3.3
django-parler-rest==2.0
django-parler==1.9.2
django-reversion==3.0.2
django-sanitized-dump==0.2.2
django==2.2.5
djangorestframework-jwt==1.11.0
djangorestframework==3.9.1
drf-oidc-auth==0.9
drf-yasg==1.16.1
easy-thumbnails==2.5
ecdsa==0.13.2
factory-boy==2.12.0
faker==1.0.2
flake8==3.6.0
freezegun==0.3.11
future==0.17.1
humanize==0.5.1
icalendar==4.0.3
idna==2.8
inflection==0.3.1
iso8601==0.1.12
isort==4.3.21
itypes==1.1.0
jinja2==2.10.1
jsonschema==2.6.0
lazy-object-proxy==1.4.2
lxml==4.3.0
markupsafe==1.1.0
mccabe==0.6.1
more-itertools==5.0.0
ntlm-auth==1.2.0
oauthlib==3.0.0
packaging==19.2
pillow==5.4.1
pip-tools==4.2.0
pluggy==0.8.1
psycopg2==2.7.6.1
py==1.7.0
pyasn1==0.4.7
pycodestyle==2.4.0
pycparser==2.19
pycryptodomex==3.7.3
pyflakes==2.0.0
pyjwkest==1.4.0
pyjwt==1.7.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pylint==2.3.1
pyparsing==2.4.2
pytest-cov==2.6.1
pytest-django==3.4.5
pytest==4.1.1
python-dateutil==2.7.5
python-docx==0.8.10
python-jose==3.0.1
python3-openid==3.1.0
pytz==2018.9
pyyaml==5.1
raven==6.10.0
requests-cache==0.4.13
requests-ntlm==1.1.0
requests-oauthlib==1.2.0
requests==2.21.0
rsa==4.0
ruamel.yaml.clib==0.1.2
ruamel.yaml==0.16.5
sentry-sdk==0.9.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
text-unidecode==1.2
typed-ast==1.4.0
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.24.3
wrapt==1.11.2
xlsxwriter==1.1.2

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.6.0        # via flake8, pytest
